feat/refacto(*): Auth system + Refacto much code
