+++ back/flag-service/.eslintrc.js
@@ -19,5 +19,14 @@ module.exports = {
     '@typescript-eslint/explicit-function-return-type': 'off',
     '@typescript-eslint/explicit-module-boundary-types': 'off',
     '@typescript-eslint/no-explicit-any': 'off',
+    '@typescript-eslint/no-unused-vars': 'error',
+    '@typescript-eslint/ban-types': [
+      "error",
+      {
+        "types": {
+          "Function": false,
+        }
+      }
+    ],
   },
 };
