+++ back/flag-service/src/main.ts
@@ -1,10 +1,3 @@
-import { NestFactory } from '@nestjs/core';
-import { FlagModule } from './flag/FlagModule';
-import { ValidationPipe } from "@nestjs/common";
+import { bootstrap } from './bootstrap';
 
-async function bootstrap() {
-  const app = await NestFactory.create(FlagModule);
-  app.useGlobalPipes(new ValidationPipe());
-  await app.listen(3000);
-}
 bootstrap();
