+++ back/flag-service/src/user/decorators/PublicDecorator.ts
@@ -1,3 +1,3 @@
 import { CustomDecorator, SetMetadata } from "@nestjs/common/decorators/core/set-metadata.decorator";
 
-export const Public = (redirect?: string): CustomDecorator => SetMetadata('public', { redirect });
+export const Public = (): CustomDecorator => SetMetadata('public', true);
