+++ back/flag-service/src/user/errors/NicknameAlreadyTakenError.ts
